requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.5
matplotlib>=3.4.3
tqdm>=4.62.3
wandb>=0.13.0
hydra-core>=1.2.0
omegaconf>=2.2.0
pyyaml>=6.0
tabulate>=0.8.9
jupyter>=1.0.0
scikit-learn>=1.0.1
pandas>=1.3.4 